Replace Icon with IconButton in PostCardHistory
The initial Icon in the PostCardHistory composable was only 24.dp in
size, which is too small for reliable interaction. A recommended
standard is a size of at least 48dp in width and height.

While we could manually add padding to this component to increase its
size, the Material component IconButton is defined to be a large enough size already.

The IconButton composable has other properties that are good for accessibility (e.g. a ripple indication to show the user that the element is clickable).